assignement9/requirements.txt
google-genai
python-dotenv
pandas
numpy
tqdm
